# api/openapi.yaml
openapi: 3.0.4
info:
  title: UnderSound Content API
  description: |
    API de contenido inspirada en Bandcamp para gestionar **álbumes, pistas, merchandising, labels y comentarios**.
    Incluye paginación, filtros, subida de archivos (multipart), acciones de publicación y favoritos.
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/escuelapolitecnicaca/contenido/1.0.0
  description: SwaggerHub API Auto Mocking
- url: https://api.undersound.example.com/v1
  description: Producción
- url: http://localhost:8081
  description: Local
security:
- oauth2:
  - content.read
  - content.write
...
paths:
...
  /merch:
    get:
      tags:
      - Merch
      summary: Listar productos de merchandising
      operationId: merchGET
      parameters:
      - name: page
        in: query
        required: false
        style: form
        explode: true
        schema:
          minimum: 1
          type: integer
          default: 1
      - name: limit
        in: query
        required: false
        style: form
        explode: true
        schema:
          maximum: 100
          minimum: 1
          type: integer
          default: 20
      - name: artistId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          description: ID del artista (puede ser número o UUID)
      - name: labelId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          description: ID del label (puede ser número o UUID)
      - name: type
        in: query
        description: Tipo de producto
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - camiseta
          - vinilo
          - cd
          - poster
          - pegatina
          - otro
      - name: availability
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - in_stock
          - out_of_stock
          - preorder
      - name: sort
        in: query
        description: Campo de ordenación
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - name
          - price
          - createdAt
      - name: order
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          default: asc
          enum:
          - asc
          - desc
      - name: q
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: Lista paginada de merchandising
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedMerchList"
      x-swagger-router-controller: Merch
    post:
      tags:
      - Merch
      summary: Crear producto de merchandising
      operationId: merchPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchCreateInput"
        required: true
      responses:
        "201":
          description: Producto creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchResponse"
      x-swagger-router-controller: Merch
  /merch/{merchId}:
    get:
      tags:
      - Merch
      summary: Detalle de producto
      operationId: merchMerchIdGET
      parameters:
      - name: merchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Detalle de producto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchResponse"
        "404":
          description: Recurso no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-swagger-router-controller: Merch
    delete:
      tags:
      - Merch
      summary: Eliminar producto
      operationId: merchMerchIdDELETE
      parameters:
      - name: merchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Eliminado
      x-swagger-router-controller: Merch
    patch:
      tags:
      - Merch
      summary: Actualizar producto
      operationId: merchMerchIdPATCH
      parameters:
      - name: merchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchUpdateInput"
        required: true
      responses:
        "200":
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchResponse"
      x-swagger-router-controller: Merch
  /merch/{merchId}/images:
    post:
      tags:
      - Uploads
      - Merch
      summary: Subir imágenes del producto
      operationId: merchMerchIdImagesPOST
      parameters:
      - name: merchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/merchId_images_body"
        required: true
      responses:
        "200":
          description: Producto con imágenes actualizadas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchResponse"
      x-swagger-router-controller: Uploads
  /merch/{merchId}/variants:
    get:
      tags:
      - Merch
      summary: Listar variantes de un producto
      operationId: merchMerchIdVariantsGET
      parameters:
      - name: merchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Lista de variantes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_1"
      x-swagger-router-controller: Merch
    post:
      tags:
      - Merch
      summary: Crear variante
      operationId: merchMerchIdVariantsPOST
      parameters:
      - name: merchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VariantCreateInput"
        required: true
      responses:
        "201":
          description: Variante creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_201"
      x-swagger-router-controller: Merch
  /merch/{merchId}/variants/{variantId}:
    delete:
      tags:
      - Merch
      summary: Eliminar variante
      operationId: merchMerchIdVariantsVariantIdDELETE
      parameters:
      - name: merchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      - name: variantId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Eliminado
      x-swagger-router-controller: Merch
    patch:
      tags:
      - Merch
      summary: Actualizar stock/precio de variante
      operationId: merchMerchIdVariantsVariantIdPATCH
      parameters:
      - name: merchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      - name: variantId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VariantUpdateInput"
        required: true
      responses:
        "200":
          description: Variante actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_201"
      x-swagger-router-controller: Merch
  /merch/{merchId}/comments:
    get:
      tags:
      - Comments
      - Merch
      summary: Listar comentarios de un producto
      operationId: merchMerchIdCommentsGET
      parameters:
      - name: merchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      - name: page
        in: query
        required: false
        style: form
        explode: true
        schema:
          minimum: 1
          type: integer
          default: 1
      - name: limit
        in: query
        required: false
        style: form
        explode: true
        schema:
          maximum: 100
          minimum: 1
          type: integer
          default: 20
      responses:
        "200":
          description: Lista paginada de comentarios
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCommentList"
      x-swagger-router-controller: Comments
    post:
      tags:
      - Comments
      - Merch
      summary: Comentar en un producto
      operationId: merchMerchIdCommentsPOST
      parameters:
      - name: merchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentCreateInput"
        required: true
      responses:
        "201":
          description: Comentario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentResponse"
        "403":
          description: "No autorizado para comentar (por ejemplo, solo si ha comprado)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-swagger-router-controller: Comments
...
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 123
      example:
        total: 123
        limit: 20
        page: 1
    Image:
      type: object
      properties:
        url:
          type: string
          format: uri
        alt:
          type: string
        width:
          type: integer
        height:
          type: integer
      example:
        alt: alt
        width: 6
        url: http://example.com/aeiou
        height: 1
    ArtistRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      example:
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
    LabelRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      example:
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
...
    MerchItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum:
          - camiseta
          - vinilo
          - cd
          - poster
          - pegatina
          - otro
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/Variant"
        label:
          $ref: "#/components/schemas/LabelRef"
        artist:
          $ref: "#/components/schemas/ArtistRef"
        availability:
          type: string
          enum:
          - in_stock
          - out_of_stock
          - preorder
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        images:
        - alt: alt
          width: 6
          url: http://example.com/aeiou
          height: 1
        - alt: alt
          width: 6
          url: http://example.com/aeiou
          height: 1
        artist:
          name: name
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        name: name
        description: description
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        variants:
        - color: color
          size: size
          price: 0.8008282
          currency: EUR
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          sku: sku
          stock: 6
        - color: color
          size: size
          price: 0.8008282
          currency: EUR
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          sku: sku
          stock: 6
        label:
          name: name
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        availability: in_stock
        type: camiseta
        updatedAt: 2000-01-23T04:56:07.000+00:00
    Variant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        color:
          type: string
        size:
          type: string
        price:
          type: number
          format: float
        currency:
          type: string
          example: EUR
        stock:
          type: integer
      example:
        color: color
        size: size
        price: 0.8008282
        currency: EUR
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        sku: sku
        stock: 6
    VariantCreateInput:
      required:
      - currency
      - price
      type: object
      properties:
        sku:
          type: string
        color:
          type: string
        size:
          type: string
        price:
          type: number
          format: float
        currency:
          type: string
        stock:
          type: integer
    VariantUpdateInput:
      type: object
      properties:
        price:
          type: number
          format: float
        currency:
          type: string
        stock:
          type: integer
    MerchCreateInput:
      required:
      - name
      - type
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum:
          - camiseta
          - vinilo
          - cd
          - poster
          - pegatina
          - otro
        artistId:
          type: string
          description: ID del artista (usuario autenticado) - puede ser número o UUID
        labelId:
          type: string
          description: ID del sello discográfico (opcional) - puede ser número o UUID
    MerchUpdateInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum:
          - camiseta
          - vinilo
          - cd
          - poster
          - pegatina
          - otro
        artistId:
          type: string
          description: ID del artista - puede ser número o UUID
        labelId:
          type: string
          description: ID del sello discográfico - puede ser número o UUID
        availability:
          type: string
          enum:
          - in_stock
          - out_of_stock
          - preorder
    PaginatedMerchList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MerchItem"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      example:
        data:
        - createdAt: 2000-01-23T04:56:07.000+00:00
          images:
          - alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          - alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          artist:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          name: name
          description: description
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          variants:
          - color: color
            size: size
            price: 0.8008282
            currency: EUR
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            sku: sku
            stock: 6
          - color: color
            size: size
            price: 0.8008282
            currency: EUR
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            sku: sku
            stock: 6
          label:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          availability: in_stock
          type: camiseta
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - createdAt: 2000-01-23T04:56:07.000+00:00
          images:
          - alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          - alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          artist:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          name: name
          description: description
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          variants:
          - color: color
            size: size
            price: 0.8008282
            currency: EUR
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            sku: sku
            stock: 6
          - color: color
            size: size
            price: 0.8008282
            currency: EUR
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            sku: sku
            stock: 6
          label:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          availability: in_stock
          type: camiseta
          updatedAt: 2000-01-23T04:56:07.000+00:00
        meta:
          total: 123
          limit: 20
          page: 1
    MerchResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/MerchItem"
      example:
        data:
          createdAt: 2000-01-23T04:56:07.000+00:00
          images:
          - alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          - alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          artist:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          name: name
          description: description
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          variants:
          - color: color
            size: size
            price: 0.8008282
            currency: EUR
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            sku: sku
            stock: 6
          - color: color
            size: size
            price: 0.8008282
            currency: EUR
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            sku: sku
            stock: 6
          label:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          availability: in_stock
          type: camiseta
          updatedAt: 2000-01-23T04:56:07.000+00:00
...
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        targetType:
          type: string
          enum:
          - album
          - track
          - merch
        targetId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        text:
          type: string
        rating:
          maximum: 5
          minimum: 1
          type: integer
        status:
          type: string
          default: visible
          enum:
          - visible
          - hidden
          - pending
        likes:
          type: integer
          default: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        targetId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        replies:
        - null
        - null
        rating: 1
        targetType: album
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        text: text
        userId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        status: visible
        likes: 6
        updatedAt: 2000-01-23T04:56:07.000+00:00
    CommentCreateInput:
      required:
      - text
      type: object
      properties:
        targetType:
          type: string
          enum:
          - album
          - track
          - merch
        targetId:
          type: string
          format: uuid
        text:
          type: string
        rating:
          maximum: 5
          minimum: 1
          type: integer
    PaginatedCommentList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      example:
        data:
        - createdAt: 2000-01-23T04:56:07.000+00:00
          targetId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          replies:
          - null
          - null
          rating: 1
          targetType: album
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          text: text
          userId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: visible
          likes: 6
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - createdAt: 2000-01-23T04:56:07.000+00:00
          targetId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          replies:
          - null
          - null
          rating: 1
          targetType: album
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          text: text
          userId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: visible
          likes: 6
          updatedAt: 2000-01-23T04:56:07.000+00:00
        meta:
          total: 123
          limit: 20
          page: 1
    CommentResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Comment"
      example:
        data:
          createdAt: 2000-01-23T04:56:07.000+00:00
          targetId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          replies:
          - null
          - null
          rating: 1
          targetType: album
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          text: text
          userId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: visible
          likes: 6
          updatedAt: 2000-01-23T04:56:07.000+00:00
...
    merchId_images_body:
      type: object
      properties:
        files:
          type: array
          items:
            type: string
            format: binary
        alt:
          type: string
    inline_response_200_1:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Variant"
      example:
        data:
        - color: color
          size: size
          price: 0.8008282
          currency: EUR
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          sku: sku
          stock: 6
        - color: color
          size: size
          price: 0.8008282
          currency: EUR
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          sku: sku
          stock: 6
    inline_response_201:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Variant"
      example:
        data:
          color: color
          size: size
          price: 0.8008282
          currency: EUR
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          sku: sku
          stock: 6
...
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.undersound.example.com/authorize
          tokenUrl: https://auth.undersound.example.com/token
          scopes:
            content.read: Lectura de contenido
            content.write: Escritura de contenido
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT